chore: clean up scripts and codes
