--- _templates/index.html
@@ -1,118 +1,0 @@
-<html>
-
-<head>
-  <title>{{package.name}}</title>
-  <meta charset="utf-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <link rel="apple-touch-icon" sizes="180x180" href="https://static.bndy.net/apple-touch-icon.png" />
-  <link rel="icon" type="image/png" sizes="32x32" href="https://static.bndy.net/favicon-32x32.png" />
-  <link rel="icon" type="image/png" sizes="16x16" href="https://static.bndy.net/favicon-16x16.png" />
-  <link rel="mask-icon" href="https://static.bndy.net/safari-pinned-tab.svg" color="#5bbad5" />
-  <link rel="stylesheet" href="style.css" />
-
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" />
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.1/jquery.slim.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-
-  <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-  <link rel="stylesheet" href="https://unpkg.com/github-markdown-css/github-markdown.css" />
-  <script src="https://unpkg.com/showdown/dist/showdown.min.js"></script>
-  <!-- dev -->
-  <link href="../dist/{{package.subname}}.css" rel="stylesheet" type="text/css" />
-  <script src="../dist/{{package.subname}}.umd.js"></script>
-  <!-- /dev -->
-
-  <!-- prod
-        <link href="https://unpkg.com/{{package.name}}/dist/{{package.subname}}.css" rel="stylesheet" type="text/css" />
-        <script src="https://unpkg.com/{{package.name}}/dist/{{package.subname}}.umd.js"></script>
-        -->
-</head>
-
-<body>
-  <a href="{{package.repository.url}}" class="forkme">
-    <img width="149" height="149"
-      src="https://github.blog/wp-content/uploads/2008/12/forkme_right_red_aa0000.png?resize=149%2C149"
-      class="attachment-full size-full" alt="Fork me on GitHub" data-recalc-dims="1">
-  </a>
-  <header>
-    <nav class="navbar fixed-top navbar-expand-lg navbar-light bg-light">
-      <a class="navbar-brand" href="#">{{package.name}}</a>
-      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent"
-        aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-
-      <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav mr-auto">
-          <li class="nav-item">
-            <!-- dev -->
-            <a class="nav-link" href="../docs/api" target="_blank">API Documentation</a>
-            <!-- /dev -->
-            <!-- prod
-                        <a class="nav-link" href="../api" target="_blank">API Documentation</a>
-                        -->
-          </li>
-          <li class="nav-item">
-            <!-- dev -->
-            <a class="nav-link" href="../coverage/lcov-report/" target="_blank">Coverage Report</a>
-            <!-- /dev -->
-
-            <!-- prod
-                        <a class="nav-link" href="../coverage-report" target="_blank">Coverage Report</a>
-                        -->
-          </li>
-        </ul>
-      </div>
-    </nav>
-  </header>
-
-  <aside>
-    <div id="sub-navs" class="list-group">
-      <a class="list-group-item list-group-item-action" href="#readme">Read me</a>
-      <a class="list-group-item list-group-item-action" href="#example">Example</a>
-      <a class="list-group-item list-group-item-action" href="#changelog">Changelog</a>
-    </div>
-  </aside>
-
-  <main data-spy="scroll" data-target="#sub-navs" data-offset="0">
-    <div class="jumbotron">
-      <h1 class="display-4">Hello, world!</h1>
-      <p class="lead">
-        This is just a template, you can start your site here for this project which will be published to
-        <b>gh-pages</b> with
-        <!-- dev -->
-        <a href="../docs/api">API documentation</a> and <a href="../coverage/lcov-report/">coverage report</a>.
-        <!-- /dev -->
-
-        <!-- prod
-                    <a href="../api">API documentation</a> and <a href="../coverage-report">coverage report</a>.
-                    -->
-      </p>
-      <hr class="my-4" />
-      <p>Try to run <code>npm run prebuild && npm run build && npm run predeploy-gh-pages</code> to view the final
-        <b>docs</b> directory for <b>gh-pages</b>.</p>
-      <a class="btn btn-primary btn-lg" href="{{package.repository.url}}" role="button">
-        Fork me on GitHub
-      </a>
-    </div>
-
-    <section id="readme" class="markdown-body"></section>
-
-    <section id="example">
-      <h1>Example</h1>
-      <p>TODO...</p>
-    </section>
-
-    <section id="changelog">
-      <h1>Changelog</h1>
-      <div id="changelog-body" class="markdown-body">
-        <em>Generated by CI automatically.</em>
-      </div>
-    </section>
-  </main>
-  <script type="text/javascript" src="./script.js"></script>
-</body>
-
-</html>
