+++ package.json
@@ -17,15 +17,12 @@
     "test": "NODE_OPTIONS=--experimental-vm-modules jest --coverage",
     "watch": "rollup --config ./rollup.config.ts --configPlugin typescript --bundleConfigAsCjs -w",
     "start": "npm run dev",
-    "site-serv": "npm run docs && npm run test && lite-server",
     "dev": "concurrently --kill-others \"npm run watch\" \"npm run website:start\"",
     "website:start": "npm run build && npm start --prefix website",
-    "website:prepare": "ts-node tools/gh-pages-prepare-beta",
+    "website:prepare": "ts-node tools/gh-pages-prepare",
     "website:build": "npm run build && npm i --prefix website && npm run website:prepare && npm run build --prefix website && npm run api",
     "test:watch": "jest --coverage --watch",
     "test:prod": "npm run lint && npm run test -- --no-cache",
-    "predeploy-gh-pages": "ts-node tools/gh-pages-prepare",
-    "deploy-gh-pages": "npm run predeploy-gh-pages && ts-node tools/gh-pages-publish",
     "report-coverage": "cat ./coverage/lcov.info | coveralls",
     "semantic-release": "semantic-release",
     "precommit": "lint-staged",
@@ -95,7 +92,6 @@
     "jsdom": "^22.1.0",
     "jsdom-global": "^3.0.2",
     "lint-staged": "^8.1.0",
-    "lite-server": "^2.5.4",
     "organize-imports-cli": "^0.8.0",
     "postcss": "^7.0.14",
     "prettier": "^2.1.2",
