+++ src/core/index.ts
@@ -1,4 +1,3 @@
-export * from './chart-element';
 export * from './constants';
 export * from './settings';
 export * from './theme';
