+++ tools/gh-pages-prepare.ts
@@ -3,32 +3,27 @@ const sh = require('shelljs');
 
 sh.echo('⚑ gh-pages preparing...');
 
-sh.mkdir('-p', './docs');
-sh.cp('README.md', './docs/README.md');
-sh.cp('CHANGELOG.md', './docs/CHANGELOG.md');
+const fileGettingStarted = './website/docs/tutorials/getting-started.md';
+sh.cp('README.md', fileGettingStarted);
+let data = fs.readFileSync(fileGettingStarted, 'utf8');
+data = data.replace(
+  '# @momentum-design/widgets',
+  `---
+sidebar_position: 1
+---
 
-sh.rm('-rf', './docs/coverage-report');
-sh.cp('-R', './coverage/lcov-report', './docs/coverage-report');
+# Getting Started`,
+);
+fs.writeFileSync(fileGettingStarted, data);
 
-sh.rm('-rf', './docs/site');
-sh.cp('-R', './site', './docs/site');
-sh.cd('./docs/site');
-sh.ls('*.html').forEach((file: string) => {
-  let data = fs.readFileSync(file, 'utf8');
-  // remove <!-- dev --> ... <!-- /dev --> lines
-  data = data.replace(/<!--\s*dev\s*-->[\s\S]*?<!--\s*\/dev\s*-->/gi, '');
-  // uncomment <!-- prod ... -->
-  data = data.replace(/<\!--\s*prod\s*([\s\S]*?)-->/gi, '$1');
-  fs.writeFileSync(file, data, (werr: any) => {
-    if (werr) {
-      throw werr;
-    }
-  });
-});
-sh.cd('../../');
 
-sh.cd('./docs');
-fs.writeFileSync('index.html', '<script>location.href="site";</script>');
-sh.cd('../');
+const fileChangelog = `./website/blog/CHANGELOG.md`;
+sh.cp('CHANGELOG.md', fileChangelog);
+data = fs.readFileSync(fileChangelog, 'utf8');
+data = data.replace(/(#+\s)/gi, '#$1');
+data = `# Changelog
+
+` + data;
+fs.writeFileSync(fileChangelog, data);
 
 sh.echo(`✔ done at ${new Date().toISOString()}`);
